Cleanup console logs and extract impressions func.
