+++ assets/js/wpjm-stats.js
@@ -3,91 +3,11 @@
 import domReady from '@wordpress/dom-ready';
 import { createHooks } from '@wordpress/hooks';
 import {
-	debounce,
-	filterZeroes,
-	findIdInClassNames,
-} from './stats/utils';
-import {
 	setUniques,
 	checkUniqueRecordedToday,
 	checkUnique,
 } from './stats/unique';
-import {
-	waitForSelector,
-} from './stats/observers';
-
-function createStatsQueue() {
-	const alreadySent = {};
-	let queue = [];
-
-	const logThem = debounce( function ( listingIds ) {
-		const stats = listingIds.map( function ( id ) {
-			alreadySent[id] = true;
-			return { name: 'job_listing_impressions', post_id: id };
-		});
-		// queue = [];
-		return wpjmLogStats( stats ).finally( function () {
-			queue = [];
-		});
-	}, 1000 );
-
-	return {
-		queueListingImpressionStats: function ( listingIds ) {
-			listingIds.forEach( function (listingId ) {
-				if ( ! alreadySent[listingId] ) {
-					queue.push(listingId);
-				}
-			} );
-			logThem( queue );
-
-		}
-	};
-}
-
-function observeForVisibility( jobListingContainer, jobListingElement, visibleCallback, alreadyViewedListings ) {
-	const options = {
-		root: null,
-		rootMargin: "0px",
-		threshold: 1.0,
-	};
-
-	const observer = new IntersectionObserver(function ( entries ) {
-		entries.forEach(function ( entry ) {
-			if ( entry.isIntersecting && entry.intersectionRatio > 0.99 ) {
-				const node = entry.target;
-				if ( 1 === node.nodeType && node.classList.contains( 'job_listing' ) ) {
-					const nodeId = findIdInClassNames( node );
-					if ( nodeId > 0 && ! alreadyViewedListings[nodeId] ) {
-						alreadyViewedListings[nodeId] = true;
-						visibleCallback( node );
-					}
-				}
-				node.classList.add( 'viewed' );
-				observer.unobserve( node );
-			}
-		} );
-	}, options );
-
-	observer.observe( jobListingElement );
-}
-
-function waitForNextVisibleListing( listingVisibleCallback ) {
-	const jobListingsContainer = document.querySelector( 'ul.job_listings' );
-	const config = { childList: true };
-	const alreadyViewed = {};
-
-	const observer = new MutationObserver(function ( mutations ) {
-		mutations.forEach(function ( mutation ) {
-			mutation.addedNodes.forEach( function ( node ) {
-				if ( 1 === node.nodeType && node.classList.contains( 'job_listing' ) && ! node.classList.contains( 'viewed' ) ) {
-					observeForVisibility( jobListingsContainer, node, listingVisibleCallback, alreadyViewed );
-				}
-			} );
-		} );
-	} );
-
-	observer.observe( jobListingsContainer, config );
-}
+import { initListingImpression } from "./stats/impressions";
 
 const WPJMStats =  {
 	init: function ( statsToRecord ) {
@@ -119,10 +39,8 @@ const WPJMStats =  {
 	},
 
 	hookStatsForTrigger: function ( statsByTrigger, triggerName ) {
-		console.log( 'hookStatsForTrigger' );
 		const statsToRecord    = [];
 		const stats            = statsByTrigger[triggerName] || [];
-		const events           = {};
 		const statsByType      = {};
 
 		stats.forEach( function ( statToRecord ) {
@@ -140,7 +58,6 @@ const WPJMStats =  {
 		}, 10 );
 
 		Object.keys( statsByType ).forEach( function ( type ) {
-			console.log( 'stats by type', type );
 			WPJMStats.types[type] && WPJMStats.types[type]( statsByType[type] );
 		} );
 	},
@@ -148,11 +65,9 @@ const WPJMStats =  {
 	hooks: createHooks(),
 	types: {
 		pageLoad: function ( stats ) {
-			console.log( 'pageLoad init script.' );
 			// This does not need to do anything special.
 		},
 		domEvent: function ( stats ) {
-			console.log( 'pageLoad init script.' );
 			const events = {};
 			stats.forEach( function ( statToRecord ) {
 				const triggerName = statToRecord.trigger;
@@ -171,24 +86,7 @@ const WPJMStats =  {
 				}
 			} );
 		},
-		// New style of declaration, a stat that relies on calling a custom js func.
-		initListingImpression: function ( stats ) {
-			console.log( 'initListingImpression init script.' );
-			const debouncedSender = createStatsQueue();
-			waitForSelector( 'li.job_listing' ).then( function () {
-				const allVisibleListings = document.querySelectorAll('li.job_listing');
-				const initialListingIds = filterZeroes( [...allVisibleListings].map( function ( elem ) {
-					return findIdInClassNames( elem );
-				} ) );
-				debouncedSender.queueListingImpressionStats( initialListingIds );
-
-				waitForNextVisibleListing( function ( elem ) {
-					const maybeId = findIdInClassNames( elem );
-					maybeId > 0 && debouncedSender.queueListingImpressionStats( [ maybeId ] );
-				} );
-
-			} );
-		},
+		initListingImpression
 
 	},
 	initCallbacks: []
